Creación de controller pero ahora apuntando a JPA, creación de clase jpa repository, adición de jpa al pom, anotación de propiedades en el modelo de todo, creación de sql que se corre en cada ejecución del proyecto

+++ ws-spring-boot-master/react-restful-ws/src/main/java/com/todo/rest/webservices/services/TodoHardcodedService.java
@@ -12,7 +12,7 @@ import com.todo.rest.webservices.models.Todo;
 public class TodoHardcodedService {
 
 	private static List<Todo> todos = new ArrayList<>();
-	private static int idCounter = 0;
+	private static Long idCounter = 0L;
 
 	static {
 		todos.add(new Todo(++idCounter, "Roman", "Learn to dance", new Date(), false));
